+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Full support for `TextureSlicer` parsing inside `NodeImage`.
 - `BoxShadow` and `TextShadow` are interpolable.
 - Support for providing `unset` in any css property.
+- Immutable components supported has been added.
+- New example showing how to spawn sound effects using css.
 
 ### Changed
 - Most changes are internal refactorings that only affect direct usage of internal APIs.